Improve bundling, esp. for compatibility with Node
Create a CommonJS (cjs) artifact for use with require() and surface it
as the primary package export.

Specify that the UMD build should be used for browsers (and unpkg, which
previously worked by referencing `main`).

Specify that the ESM artifact should be used when loading this package
as a module.